Mock TwitchIrcServer & Channel in mocks & make getSevenTvBadges mockable
