+++ mocks/include/mocks/EmptyApplication.hpp
@@ -74,6 +74,12 @@ public:
         return nullptr;
     }
 
+    SeventvBadges *getSeventvBadges() override
+    {
+        assert(!"getSeventvBadges was called without being initialized");
+        return nullptr;
+    }
+
     IUserDataController *getUserData() override
     {
         return nullptr;
